+++ Examples/Resources/ADObjectEnabledState/3-ADObjectEnabledState_EnabledPrestagedClusterComputerAccount_Config.ps1
@@ -30,7 +30,7 @@ Configuration ADObjectEnabledState_EnabledPrestagedClusterComputerAccount_Config
         [Parameter(Mandatory = $true)]
         [ValidateNotNullOrEmpty()]
         [System.Management.Automation.PSCredential]
-        $DomainAdministratorCredential
+        $Credential
     )
 
     Import-DscResource -ModuleName ActiveDirectoryDsc
@@ -42,7 +42,7 @@ Configuration ADObjectEnabledState_EnabledPrestagedClusterComputerAccount_Config
         {
             Name                          = 'CLU_CNO01'
             StaticIPAddress               = '192.168.100.20/24'
-            DomainAdministratorCredential = $DomainAdministratorCredential
+            DomainAdministratorCredential = $Credential
         }
 
         ADObjectEnabledState 'EnforceEnabledPropertyToEnabled'